Ticket choosing flow

+++ src/models/setting_store/setting.go
@@ -4,12 +4,12 @@ import "time"
 
 type Setting struct {
 	ID int `json:"id"`
-	TicketPrizeE5 int `json:"ticket_prize_e5"`
+	TicketPrizeE5 int64 `json:"ticket_prize_e5"`
 	PayoutPercent int `json:"payout_percent"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
 
 type CreateSettingRequest struct {
-	TicketPrizeE5 int `json:"ticket_prize_e5"`
+	TicketPrizeE5 int64 `json:"ticket_prize_e5"`
 	PayoutPercent int `json:"payout_percent"`
 }
